having issues with pytest
